update(env): mv environ -> variables
