+++ srcs/destroy.c
@@ -1,5 +1,6 @@
 #include <readline/readline.h>
 #include "minishell.h"
+#include "ms_var.h"
 #include "ft_mem.h"
 
 void	destroy(t_context context)
