+++ srcs/init.c
@@ -1,7 +1,7 @@
 #include <stdlib.h>
 #include <unistd.h>
 #include "minishell.h"
-#include "ms_builtin.h"
+#include "ms_var.h"
 #include "ft_string.h"
 
 static void	set_context_initial_value(t_context *context)
