+++ srcs/variables/clear.c
@@ -1,4 +1,5 @@
 #include "minishell.h"
+#include "ms_var.h"
 #include "ft_hash.h"
 
 void	env_clear(t_env *env)
