+++ srcs/variables/declare_arg.c
@@ -1,5 +1,6 @@
 #include "minishell.h"
 #include "ms_builtin.h"
+#include "ms_var.h"
 #include "ft_mem.h"
 #include "ft_string.h"
 #include "ft_sys.h"
