+++ srcs/variables/dup_env_key.c
@@ -1,4 +1,5 @@
 #include "minishell.h"
+#include "ms_var.h"
 #include "ft_string.h"
 
 static bool	is_add_operator(const char c)
