+++ srcs/variables/env_separate_variables.c
@@ -1,8 +1,7 @@
 #include "minishell.h"
-#include "ms_builtin.h"
+#include "ms_var.h"
 #include "ft_mem.h"
 #include "ft_string.h"
-#include "ft_sys.h"
 
 static bool	is_only_key(const char c)
 {
