+++ srcs/variables/set_default_old_pwd.c
@@ -1,7 +1,6 @@
-#include <dirent.h>
 #include <errno.h>
 #include "minishell.h"
-#include "ms_builtin.h"
+#include "ms_var.h"
 #include "ft_mem.h"
 #include "ft_string.h"
