+++ srcs/variables/set_default_pwd.c
@@ -1,7 +1,6 @@
 #include "minishell.h"
-#include "ms_builtin.h"
+#include "ms_var.h"
 #include "ft_mem.h"
-#include "ft_string.h"
 
 void	set_default_pwd(t_env *env)
 {
